Modularize the code. This helps group 'components' together, and refactor classmethod-only classes to modules.
