+++ reactives/tests/test_collections.py
@@ -1,397 +1,10 @@
-import gc
 import sys
 import unittest
 from unittest import TestCase
 
-from reactives import ReactorController, UnsupportedReactive, reactive, ReactiveDict, ReactiveList, \
-    _InstanceReactorController
-from reactives.tests import assert_reactor_called, assert_in_scope, assert_not_reactor_called, assert_is_reactive, \
-    assert_scope_empty
-
-
-class ReactiveTest(TestCase):
-    def test_with_unsupported_reactive(self) -> None:
-        with self.assertRaises(UnsupportedReactive):
-            reactive(999)
-
-
-class ReactorControllerTest(TestCase):
-    class _Reactive:
-        def __init__(self):
-            self.react = ReactorController()
-
-    def test_react_without_reactors(self) -> None:
-        sut = ReactorController()
-        sut.trigger()
-
-    def test_react_with_reactor(self) -> None:
-        sut = ReactorController()
-        with assert_reactor_called() as reactor:
-            sut.react(reactor)
-            sut.trigger()
-
-    def test_react_with_diamond_reactors(self) -> None:
-        sut = ReactorController()
-        with assert_reactor_called() as final_reactor:
-            intermediate_reactive_1 = self._Reactive()
-            intermediate_reactive_2 = self._Reactive()
-            intermediate_reactive_1.react(final_reactor)
-            intermediate_reactive_2.react(final_reactor)
-            sut.react(intermediate_reactive_1)
-            sut.react(intermediate_reactive_2)
-            sut.trigger()
-
-    def test_react_with_intermediate_diamond_reactors(self) -> None:
-        order_tracker = []
-        r_a = self._Reactive()
-        r_a.react(lambda: order_tracker.append('r_a'))
-        r_b = self._Reactive()
-        r_b.react(lambda: order_tracker.append('r_b'))
-        r_c = self._Reactive()
-        r_c.react(lambda: order_tracker.append('r_c'))
-        r_d = self._Reactive()
-        r_d.react(lambda: order_tracker.append('r_d'))
-        r_da = self._Reactive()
-        r_da.react(lambda: order_tracker.append('r_da'))
-        r_e = self._Reactive()
-        r_e.react(lambda: order_tracker.append('r_e'))
-
-        r_a.react(r_b)
-        r_b.react(r_c)
-        r_b.react(r_d)
-
-        # This is the intermediate trigger we are asserting is consolidated into the reactor chain.
-        def f_ca():
-            order_tracker.append('f_ca')
-            r_e.react.trigger()
-        r_c.react(f_ca)
-        r_d.react(r_da)
-        r_da.react(r_e)
-
-        r_a.react.trigger()
-        self.assertEquals(['r_a', 'r_b', 'r_c', 'f_ca',
-                           'r_d', 'r_da', 'r_e'], order_tracker)
-
-    def test_react_using_shortcut_with_reactor(self) -> None:
-        sut = ReactorController()
-        with assert_reactor_called() as reactor:
-            sut(reactor)
-            sut.trigger()
-
-    def test_shutdown(self) -> None:
-        sut = ReactorController()
-        with assert_reactor_called() as reactor:
-            sut.react(reactor)
-            sut.trigger()
-            sut.shutdown(reactor)
-            sut.trigger()
-
-    def test_react_weakref(self) -> None:
-        sut = ReactorController()
-        reactor = assert_not_reactor_called()
-        sut.react_weakref(reactor)
-        del reactor
-        gc.collect()
-        sut.trigger()
-
-    def test_shutdown_weakref(self) -> None:
-        sut = ReactorController()
-        with assert_reactor_called() as reactor:
-            sut.react_weakref(reactor)
-            sut.trigger()
-            sut.shutdown_weakref(reactor)
-            sut.trigger()
-
-    def test_suspend(self) -> None:
-        sut = ReactorController()
-        with assert_not_reactor_called() as reactor:
-            sut.react_weakref(reactor)
-            with ReactorController.suspend():
-                sut.trigger()
-
-
-class InstanceReactorControllerTest(TestCase):
-    def test_getattr_with_reactive_attribute(self) -> None:
-        @reactive
-        class Subject:
-            @reactive
-            def subject(self):
-                pass
-        sut = _InstanceReactorController(Subject())
-        assert_is_reactive(sut.getattr('subject'))
-
-    def test_getattr_with_non_existent_reactive_attribute(self) -> None:
-        @reactive
-        class Subject:
-            def subject(self):
-                pass
-        sut = _InstanceReactorController(Subject())
-        with self.assertRaises(AttributeError):
-            sut.getattr('subject')
-
-
-class ReactiveFunctionTest(TestCase):
-    def test_without_reactors(self) -> None:
-        @reactive
-        def subject():
-            subject.tracker.append(True)
-        subject.tracker = []
-        subject()
-        self.assertEquals([True], subject.tracker)
-
-    def test_with_reactor(self) -> None:
-        @reactive
-        def subject():
-            raise AssertionError('This function should not have been called.')
-
-        with assert_reactor_called(subject):
-            subject.react.trigger()
-
-    def test_with_reactor_and_dependency(self) -> None:
-        @reactive
-        def dependency():
-            pass
-
-        @reactive
-        def subject():
-            if not subject.called:
-                subject.called = True
-                dependency()
-        subject.called = False
-        with assert_reactor_called(subject):
-            # Call the reactive for the first time. This should result in dependency() being autowired.
-            subject()
-            # dependency() being autowired should cause the reactor to be called.
-            dependency.react.trigger()
-
-            # Call the reactive again. This should result in dependency() being ignored and not to be autowired again.
-            subject()
-            # dependency() no longer being autowired should not cause the reactor to be called.
-            dependency.react.trigger()
-
-    def test_on_trigger_call(self):
-        @reactive(on_trigger_call=True)
-        def subject():
-            subject.tracker.append(True)
-        subject.tracker = []
-        subject.react.trigger()
-        self.assertEquals([True], subject.tracker)
-
-    def test_on_trigger_call_as_instance_method(self):
-        @reactive
-        class Subject:
-            def __init__(self):
-                self.tracker = []
-
-            @reactive(on_trigger_call=True)
-            def subject(self):
-                self.tracker.append(True)
-        subject = Subject()
-        subject.react.getattr('subject').react.trigger()
-        self.assertEquals([True], subject.tracker)
-
-
-class ReactiveInstanceTest(TestCase):
-    def test_instance_trigger_without_reactors(self) -> None:
-        @reactive
-        class Subject:
-            pass
-        Subject().react.trigger()
-
-    def test_instance_trigger_with_instance_reactor(self) -> None:
-        @reactive
-        class Subject:
-            pass
-        subject = Subject()
-        with assert_reactor_called(subject):
-            subject.react.trigger()
-
-    def test_instance_trigger_with_instance_attribute_reactor(self) -> None:
-        @reactive
-        class Subject:
-            @reactive
-            def subject(self) -> None:
-                pass
-        subject = Subject()
-        with assert_not_reactor_called(subject.react.getattr('subject')):
-            Subject().react.trigger()
-
-    def test_instance_attribute_trigger_without_reactors(self) -> None:
-        @reactive
-        class Subject:
-            @reactive
-            def subject(self) -> None:
-                pass
-        Subject().react.getattr('subject').react.trigger()
-
-    def test_instance_attribute_trigger_with_instance_reactor(self) -> None:
-        @reactive
-        class Subject:
-            @reactive
-            def subject(self) -> None:
-                pass
-        subject = Subject()
-        with assert_reactor_called(subject):
-            subject.react.getattr('subject').react.trigger()
-
-    def test_instance_attribute_trigger_with_instance_attribute_reactor(self) -> None:
-        @reactive
-        class Subject:
-            @reactive
-            def subject(self) -> None:
-                pass
-        subject = Subject()
-        with assert_reactor_called(subject.react.getattr('subject')):
-            subject.react.getattr('subject').react.trigger()
-
-
-class ReactivePropertyTest(TestCase):
-    def test_without_fget(self) -> None:
-        with self.assertRaises(UnsupportedReactive):
-            reactive(property())
-
-    def test_fget(self) -> None:
-        @reactive
-        def dependency():
-            pass
-
-        @reactive
-        class Subject:
-            def __init__(self):
-                self._subject_called = False
-
-            @reactive
-            @property
-            def subject(self):
-                if not self._subject_called:
-                    self._subject_called = True
-                    return dependency()
-
-            @reactive
-            @property
-            def subject2(self):
-                if not self._subject_called:
-                    self._subject_called = True
-                    return dependency()
-
-        subject = Subject()
-
-        with assert_reactor_called(subject):
-            with assert_in_scope(subject.react.getattr('subject')):
-                # Call the reactive for the first time. This should result in dependency() being autowired.
-                subject.subject
-
-            # dependency() being autowired should cause the reactor to be called.
-            dependency.react.trigger()
-
-            # Call the reactive again. This should result in dependency() being ignored and not to be autowired again.
-            with assert_in_scope(subject.react.getattr('subject')):
-                subject.subject
-
-        # dependency() no longer being autowired should not cause the reactor to be called.
-        with assert_not_reactor_called(subject):
-            dependency.react.trigger()
-
-    def test_fset(self) -> None:
-        @reactive
-        class DependencyOne:
-            pass
-
-        @reactive
-        class DependencyTwo:
-            pass
-
-        @reactive
-        class Subject:
-            def __init__(self):
-                self._subject = 123
-
-            @reactive
-            @property
-            def subject(self):
-                return self._subject
-
-            @subject.setter
-            def subject(self, subject) -> None:
-                self._subject = subject
-
-        subject = Subject()
-        dependency_one = DependencyOne()
-        dependency_two = DependencyTwo()
-
-        # Setting dependency_one should cause the reactor to be called.
-        with assert_reactor_called(subject.react.getattr('subject')):
-            subject.subject = dependency_one
-            self.assertEquals(dependency_one, subject._subject)
-
-        # dependency_one being autowired should cause the reactor to be called.
-        with assert_reactor_called(subject.react.getattr('subject')):
-            dependency_one.react.trigger()
-
-        # Setting dependency_two should cause the reactor to be called.
-        with assert_reactor_called(subject.react.getattr('subject')):
-            subject.subject = dependency_two
-            self.assertEquals(dependency_two, subject._subject)
-
-        # dependency_one no longer being autowired should not cause the reactor to be called.
-        with assert_not_reactor_called(subject.react.getattr('subject')):
-            dependency_one.react.trigger()
-
-    def test_fdel(self) -> None:
-        @reactive
-        class Dependency:
-            pass
-
-        @reactive
-        class Subject:
-            def __init__(self):
-                self._subject = 123
-
-            @reactive
-            @property
-            def subject(self):
-                return self._subject
-
-            @subject.setter
-            def subject(self, subject) -> None:
-                self._subject = subject
-
-            @subject.deleter
-            def subject(self) -> None:
-                self._subject = None
-
-        subject = Subject()
-
-        # Even if the property's setter and getter weren't called, deletion should cause the reactor to be called.
-        with assert_reactor_called(subject.react.getattr('subject')):
-            del subject.subject
-        self.assertIsNone(subject._subject)
-
-        dependency = Dependency()
-
-        # Setting dependency will autowire it.
-        subject.subject = dependency
-
-        # dependency_one no longer being autowired should not cause the reactor to be called.
-        del subject.subject
-        subject.react.getattr('subject').react.react_weakref(
-            assert_not_reactor_called())
-        dependency.react.trigger()
-
-    def test_on_trigger(self) -> None:
-        @reactive
-        class Subject:
-            def __init__(self):
-                self._subject = 123
-
-            @reactive(on_trigger=(lambda instance: setattr(instance, '_subject', None),))
-            @property
-            def subject(self):
-                return self._subject
-
-        subject = Subject()
-        subject.react.getattr('subject').react.trigger()
-        self.assertIsNone(subject._subject)
+from reactives.collections import ReactiveList, ReactiveDict
+from reactives.reactive import reactive
+from reactives.tests import assert_scope_empty, assert_reactor_called, assert_in_scope, assert_not_reactor_called
 
 
 class ReactiveDictTest(TestCase):
